requirements.txt
absl-py==1.3.0
altair==5.2.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
cleo==2.1.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.0.6
crashtest==0.4.1
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.12.2
gast==0.5.4
gdown==4.7.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.14.0
google-auth-oauthlib==0.7.1
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.10.0
hub-sdk==0.0.3
humanfriendly==10.0
idna==3.4
imageio==2.22.3
imgviz==1.7.5
importlib-metadata==7.0.1
installer==0.7.0
ipykernel==6.29.4
ipython==8.25.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.3
keras==2.15.0
keyring==24.3.0
kiwisolver==1.4.4
labelme==5.4.1
libclang==16.0.6
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
natsort==8.4.0
nest-asyncio==1.6.0
networkx==2.6.3
numpy==1.25.2
oauthlib==3.2.2
onnxruntime==1.16.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==2.1.4
parso==0.8.4
pexpect==4.9.0
Pillow==9.3.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt_toolkit==3.0.46
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.0.9
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==26.0.3
QtPy==2.4.1
rapidfuzz==3.6.1
referencing==0.32.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.13.1
Shapely==1.8.5.post1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
text-unidecode==1.3
thop==0.1.1.post2209072238
tifffile==2021.11.2
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trove-classifiers==2024.1.8
typing_extensions==4.12.1
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.2
urllib3==1.26.13
validators==0.22.0
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
xattr==0.10.1
zipp==3.17.0
